Prepare UI to accept tags
