+++ src/main/java/seedu/address/ui/PersonCard.java
@@ -1,17 +1,14 @@
 package seedu.address.ui;
 
-import java.util.Comparator;
-import java.util.Set;
-import java.util.stream.Stream;
-
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
 import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
 import seedu.address.model.person.Person;
-import seedu.address.model.tag.ChildTag;
-import seedu.address.model.tag.Tag;
+
+import static seedu.address.ui.UiUtil.generateTagLabel;
+import static seedu.address.ui.UiUtil.streamTags;
 
 /**
  * An UI component that displays information of a {@code Person}.
@@ -57,44 +54,6 @@ public class PersonCard extends UiPart<Region> {
         address.setText(person.getAddress().value);
         email.setText(person.getEmail().value);
         streamTags(person.getTags()).forEach(tag -> tags.getChildren().add(generateTagLabel(tag)));
-    }
-
-    /**
-     * Generates a label of different colors based on whether the tag is a ChildTag.
-     * @param tag the tag to create a label from.
-     * @return Label with the tagName as text and different colors based on the tag type.
-     */
-    private Label generateTagLabel(Tag tag) {
-        Label label = new Label(tag.tagName);
-        if (tag instanceof ChildTag) {
-            label.setStyle("-fx-background-color: #ff5050");
-        }
-        return label;
-    }
-
-    private Stream<Tag> streamTags(Set<Tag> tagSet) {
-        return tagSet.stream()
-                .sorted(new TagComparator());
-    }
-
-    private class TagComparator implements Comparator<Tag> {
-
-        @Override
-        public int compare(Tag tag1, Tag tag2) {
-            if (tag1 instanceof ChildTag) {
-                if (tag2 instanceof ChildTag) {
-                    return String.CASE_INSENSITIVE_ORDER.compare(tag1.tagName, tag2.tagName);
-                } else {
-                    return -1;
-                }
-            } else {
-                if (tag2 instanceof ChildTag) {
-                    return 1;
-                } else {
-                    return String.CASE_INSENSITIVE_ORDER.compare(tag1.tagName, tag2.tagName);
-                }
-            }
-        }
     }
 
     @Override
